Rename omx to open_manipulator_x
