+++ open_manipulator_bringup/config/open_manipulator_x/hardware_controller_manager.yaml
