+++ open_manipulator_bringup/config/open_manipulator_x/initial_positions.yaml
