+++ open_manipulator_bringup/setup.py
@@ -14,14 +14,14 @@ setup(
         ('share/ament_index/resource_index/packages', ['resource/' + package_name]),
         ('share/' + package_name, ['package.xml']),
         (os.path.join('share', package_name, 'launch'), glob('launch/*.py')),
-        (os.path.join('share', package_name, 'config/omx'), glob('config/omx/*')),
+        (os.path.join('share', package_name, 'config/open_manipulator_x'), glob('config/open_manipulator_x/*')),
         (os.path.join('share', package_name, 'config/omy_3m'), glob('config/omy_3m/*')),
         (os.path.join('share', package_name, 'config/omy_f3m'), glob('config/omy_f3m/*')),
         (os.path.join('share', package_name, 'config/omy_f3m_follower_ai'),
             glob('config/omy_f3m_follower_ai/*')),
         (os.path.join('share', package_name, 'config/omy_l100_leader_ai'),
             glob('config/omy_l100_leader_ai/*')),
-        (os.path.join('share', package_name, 'config/omx'), glob('config/omx/*')),
+        (os.path.join('share', package_name, 'config/open_manipulator_x'), glob('config/open_manipulator_x/*')),
         (os.path.join('share', package_name, 'worlds'), glob('worlds/*')),
         (os.path.join('share', package_name), ['open-manipulator-cdc.rules']),
