+++ open_manipulator_description/meshes/open_manipulator_x/gripper_left_palm.stl
