+++ open_manipulator_description/meshes/open_manipulator_x/gripper_right_palm.stl
