+++ open_manipulator_description/meshes/open_manipulator_x/link1.stl
