+++ open_manipulator_description/meshes/open_manipulator_x/link2.stl
