+++ open_manipulator_description/urdf/open_manipulator_x/open_manipulator_x.urdf.xacro
@@ -1,15 +1,15 @@
 <?xml version="1.0"?>
-<robot name="omx" xmlns:xacro="http://ros.org/wiki/xacro">
+<robot name="open_manipulator_x" xmlns:xacro="http://ros.org/wiki/xacro">
   <xacro:arg name="prefix" default="" />
   <xacro:arg name="use_sim" default="false" />
   <xacro:arg name="use_fake_hardware" default="false" />
   <xacro:arg name="fake_sensor_commands" default="false" />
   <xacro:arg name="port_name" default="/dev/ttyUSB0" />
-  <xacro:arg name="ros2_control_type" default="omx" />
+  <xacro:arg name="ros2_control_type" default="open_manipulator_x" />
 
-  <xacro:include filename="$(find open_manipulator_description)/gazebo/omx.gazebo.xacro" />
+  <xacro:include filename="$(find open_manipulator_description)/gazebo/open_manipulator_x.gazebo.xacro" />
 
-  <xacro:include filename="$(find open_manipulator_description)/urdf/omx/omx_arm.urdf.xacro" />
+  <xacro:include filename="$(find open_manipulator_description)/urdf/open_manipulator_x/open_manipulator_x_arm.urdf.xacro" />
 
   <xacro:include filename="$(find open_manipulator_description)/ros2_control/$(arg ros2_control_type).ros2_control.xacro" />
 
@@ -21,18 +21,18 @@
     <child link="$(arg prefix)link1"/>
   </joint>
 
-  <xacro:omx prefix="$(arg prefix)" />
-  <xacro:omx_gazebo prefix="$(arg prefix)" />
+  <xacro:open_manipulator_x prefix="$(arg prefix)" />
+  <xacro:open_manipulator_x_gazebo prefix="$(arg prefix)" />
 
-  <xacro:omx_system
-    name="OMXSystem" prefix="$(arg prefix)" use_sim="$(arg use_sim)"
+  <xacro:open_manipulator_x_system
+    name="OpenManipulatorXSystem" prefix="$(arg prefix)" use_sim="$(arg use_sim)"
     use_fake_hardware="$(arg use_fake_hardware)"
     fake_sensor_commands="$(arg fake_sensor_commands)"
     port_name="$(arg port_name)"/>
 
   <gazebo>
     <plugin filename="gz_ros2_control-system" name="gz_ros2_control::GazeboSimROS2ControlPlugin">
-      <parameters>$(find open_manipulator_bringup)/config/omx/hardware_controller_manager.yaml</parameters>
+      <parameters>$(find open_manipulator_bringup)/config/open_manipulator_x/hardware_controller_manager.yaml</parameters>
     </plugin>
   </gazebo>
