+++ open_manipulator_gui/launch/open_manipulator_x_gui.launch.py
@@ -27,12 +27,12 @@ def generate_launch_description():
     csv_file_path = os.path.join(
         get_package_share_directory('open_manipulator_gui'),
         'config',
-        'omx_joint_log.csv',
+        'open_manipulator_x_joint_log.csv',
     )
 
     gui_node = Node(
         package='open_manipulator_gui',
-        executable='omx_gui_node',
+        executable='open_manipulator_x_gui_node',
         output='screen',
         parameters=[{'use_sim_time': True}, {'csv_path': csv_file_path}],
     )
