+++ open_manipulator_gui/src/open_manipulator_x/main.cpp
@@ -16,14 +16,14 @@
 
 #include <QtGui>
 #include <QApplication>
-#include "../include/omx_gui/main_window.hpp"
-#include "omx_gui/qnode.hpp"
+#include "../include/open_manipulator_x_gui/main_window.hpp"
+#include "open_manipulator_x_gui/qnode.hpp"
 
 
 int main(int argc, char ** argv)
 {
   QApplication app(argc, argv);
-  omx_gui::MainWindow w(argc, argv);
+  open_manipulator_x_gui::MainWindow w(argc, argv);
   w.show();
   app.connect(&app, SIGNAL(lastWindowClosed()), &app, SLOT(quit()));
