+++ open_manipulator_gui/ui/open_manipulator_x_main_window.ui
