+++ open_manipulator_playground/src/open_manipulator_x_hello_moveit.cpp
@@ -14,7 +14,7 @@
 //
 // Author: Sungho Woo, Wonho Yun, Woojin Wie
 
-#include "open_manipulator_playground/omx_hello_moveit.h"
+#include "open_manipulator_playground/open_manipulator_x_hello_moveit.h"
 
 #include <memory>
 #include <chrono>
